modificacion de hoja de trabajo y user

+++ bsuser/views.py
@@ -8,10 +8,9 @@ from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
 
 # Create your views here.
 
-def login(request):
-	return render(request, "login.html")
-
 def home_user(request):
+	if not request.user.is_authenticated() or request.user.is_staff:
+		raise Http404
 	#DetAna_queryset = DetalleAnalisisPadre.objects.all().order_by('-protocolo')
 	DetAna_queryset_list = DetalleAnalisisPadre.objects.distinct('protocolo')
 	
@@ -59,6 +58,8 @@ def j_solicitudanalisisid(request,id=None):
 # 	return render(request, "list_solAn.html", context)
 
 def a_solicitudanalisis(request):
+	if not request.user.is_authenticated() or request.user.is_staff:
+		raise Http404
 	form = SolicitudAnalisisForm(request.POST or None)
 	if form.is_valid():
 		instance = form.save(commit=False)
@@ -81,6 +82,8 @@ def a_solicitudanalisis(request):
 # 	return render(request, "ver_solAn.html", context)
 
 def u_solicitudanalisis(request, id=None):
+	if not request.user.is_authenticated() or request.user.is_staff:
+		raise Http404
 	instance = get_object_or_404(SolicitudAnalisis,id=id)
 	form = SolicitudAnalisisForm(request.POST or None, instance=instance)
 	if form.is_valid():
@@ -97,6 +100,8 @@ def u_solicitudanalisis(request, id=None):
 	return render(request, "alta_solAn.html", context)
 
 def d_solicitudanalisis(request, id=None):
+	if not request.user.is_authenticated() or request.user.is_staff:
+		raise Http404
 	instance = get_object_or_404(SolicitudAnalisis, id=id)
 	instance.delete()
 	return redirect("bsadmin:l_solicitudanalisis")
@@ -122,6 +127,8 @@ def j_protocoloid(request,id=None):
 # 	return render(request, "list_protocolo.html", context)
 
 def a_protocolo(request):
+	if not request.user.is_authenticated() or request.user.is_staff:
+		raise Http404
 	form = ProtocoloForm(request.POST or None)
 	if form.is_valid():
 		instance = form.save(commit=False)
@@ -136,6 +143,8 @@ def a_protocolo(request):
 	return render(request, "alta_protocolo.html", context)
 
 def u_protocolo(request, id=None):
+	if not request.user.is_authenticated() or request.user.is_staff:
+		raise Http404
 	instance = get_object_or_404(Protocolo,id=id)
 	form = ProtocoloForm(request.POST or None, instance=instance)
 	if form.is_valid():
@@ -152,6 +161,8 @@ def u_protocolo(request, id=None):
 	return render(request, "alta_protocolo.html", context)
 
 def d_protocolo(request, id=None):
+	if not request.user.is_authenticated() or request.user.is_staff:
+		raise Http404
 	instance = get_object_or_404(Protocolo, id=id)
 	instance.delete()
 	return redirect("bsadmin:l_protocolo")
@@ -177,6 +188,8 @@ def j_individuopadreid(request,id=None):
 # 	return render(request, "list_indivp.html", context)
 
 def a_individuopadre(request):
+	if not request.user.is_authenticated() or request.user.is_staff:
+		raise Http404
 	form = IndividuoPadreForm(request.POST or None)
 	if form.is_valid():
 		instance = form.save(commit=False)
@@ -199,6 +212,8 @@ def a_individuopadre(request):
 # 	return render(request, "ver_indivp.html", context)
 
 def u_individuopadre(request, id=None):
+	if not request.user.is_authenticated() or request.user.is_staff:
+		raise Http404
 	instance = get_object_or_404(IndividuoPadre,id=id)
 	form = IndividuoPadreForm(request.POST or None, instance=instance)
 	if form.is_valid():
@@ -215,6 +230,8 @@ def u_individuopadre(request, id=None):
 	return render(request, "alta_indivp.html", context)
 
 def d_individuopadre(request, id=None):
+	if not request.user.is_authenticated() or request.user.is_staff:
+		raise Http404
 	instance = get_object_or_404(IndividuoPadre, id=id)
 	instance.delete()
 	return redirect("bsadmin:l_individuopadre")
@@ -262,6 +279,8 @@ def j_individuosid(request,id=None):
 # 	return render(request, "ver_indiv.html", context)
 
 def u_individuos(request, id=None):
+	if not request.user.is_authenticated() or request.user.is_staff:
+		raise Http404
 	instance = get_object_or_404(Individuos,id=id)
 	form = IndividuosForm(request.POST or None, instance=instance)
 	if form.is_valid():
@@ -278,6 +297,8 @@ def u_individuos(request, id=None):
 	return render(request, "alta_indiv.html", context)
 
 def d_individuos(request, id=None):
+	if not request.user.is_authenticated() or request.user.is_staff:
+		raise Http404
 	instance = get_object_or_404(Individuos, id=id)
 	instance.delete()
 	return redirect("bsadmin:l_individuos")
@@ -317,6 +338,8 @@ def j_detalleanalisisid(request,id=None):
 # 	return render(request, "alta_detAn.html", context)
 
 def v_detalleanalisis(request, id=None):
+	if not request.user.is_authenticated() or request.user.is_staff:
+		raise Http404
 	instance = get_object_or_404(DetalleAnalisis, id=id)
 	context = {
 		"instance" : instance,
@@ -325,6 +348,8 @@ def v_detalleanalisis(request, id=None):
 	return render(request, "ver_detAn.html", context)
 
 def u_detalleanalisis(request, id=None):
+	if not request.user.is_authenticated() or request.user.is_staff:
+		raise Http404
 	instance = get_object_or_404(DetalleAnalisis,id=id)
 	form = DetalleAnalisisForm(request.POST or None, instance=instance)
 	if form.is_valid():
@@ -341,6 +366,8 @@ def u_detalleanalisis(request, id=None):
 	return render(request, "alta_detAn.html", context)
 
 def d_detalleanalisis(request, id=None):
+	if not request.user.is_authenticated() or request.user.is_staff:
+		raise Http404
 	instance = get_object_or_404(DetalleAnalisis, id=id)
 	instance.delete()
 	return redirect("bsadmin:l_detalleanalisis")
@@ -358,6 +385,8 @@ def j_DetalleAnalisisPadreid(request,id=None):
 	return JsonResponse(queryset, safe=False)
 
 def l_DetalleAnalisisPadre(request):
+	if not request.user.is_authenticated() or request.user.is_staff:
+		raise Http404
 	queryset = DetalleAnalisisPadre.objects.all().order_by('id')
 	context = {
 		"object_list": queryset,
@@ -366,6 +395,8 @@ def l_DetalleAnalisisPadre(request):
 	return render(request, "list_detAn.html", context)
 
 def a_DetalleAnalisisPadre(request):
+	if not request.user.is_authenticated() or request.user.is_staff:
+		raise Http404
 	form = DetalleAnalisisPadreForm(request.POST or None)
 	if form.is_valid():
 		instance = form.save(commit=False)
@@ -380,6 +411,8 @@ def a_DetalleAnalisisPadre(request):
 	return render(request, "alta_detAn.html", context)
 
 def v_DetalleAnalisisPadre(request, id=None):
+	if not request.user.is_authenticated() or request.user.is_staff:
+		raise Http404
 	instance = get_object_or_404(DetalleAnalisisPadre, id=id)
 	prot = instance.protocolo
 	queryset = DetalleAnalisisPadre.objects.all().order_by('id').filter(protocolo=prot)
@@ -403,6 +436,8 @@ def v_DetalleAnalisisPadre(request, id=None):
 	return render(request, "ver_detAn.html", context)
 
 def u_DetalleAnalisisPadre(request, id=None):
+	if not request.user.is_authenticated() or request.user.is_staff:
+		raise Http404
 	instance = get_object_or_404(DetalleAnalisisPadre,id=id)
 	# INSTANCIA_INDIVIDUOS
 	# INSTANCIA_DIAGNOSTICO
@@ -441,6 +476,8 @@ def j_eliminacionprotocoloid(request,id=None):
 	return JsonResponse(queryset, safe=False)
 
 def l_eliminacionprotocolo(request):
+	if not request.user.is_authenticated() or request.user.is_staff:
+		raise Http404
 	queryset = EliminacionProtocolo.objects.all().order_by('id')
 	query = request.GET.get("q")
 	if query:
@@ -466,6 +503,8 @@ def l_eliminacionprotocolo(request):
 	return render(request, "list_elimProt.html", context)
 
 def a_eliminacionprotocolo(request, id=None):
+	if not request.user.is_authenticated() or request.user.is_staff:
+		raise Http404
 	instance = get_object_or_404(DetalleAnalisisPadre, id=id)
 	form = EliminacionProtocoloForm(request.POST or None)
 	if form.is_valid():
@@ -486,6 +525,8 @@ def a_eliminacionprotocolo(request, id=None):
 	return render(request, "alta_elimProt.html", context)
 
 def v_eliminacionprotocolo(request, id=None):
+	if not request.user.is_authenticated() or request.user.is_staff:
+		raise Http404
 	instance = get_object_or_404(EliminacionProtocolo, id=id)
 	context = {
 		"instance" : instance,
@@ -494,6 +535,8 @@ def v_eliminacionprotocolo(request, id=None):
 	return render(request, "ver_elimProt.html", context)
 
 def u_eliminacionprotocolo(request, id=None):
+	if not request.user.is_authenticated() or request.user.is_staff:
+		raise Http404
 	instance_ep = get_object_or_404(EliminacionProtocolo,id=id)
 	dap_list = DetalleAnalisisPadre.objects.filter(protocolo=instance_ep.protocolo).distinct()
 	for dap in dap_list:
@@ -514,11 +557,15 @@ def u_eliminacionprotocolo(request, id=None):
 	return render(request, "alta_elimProt.html", context)
 
 def d_eliminacionprotocolo(request, id=None):
+	if not request.user.is_authenticated() or request.user.is_staff:
+		raise Http404
 	instance = get_object_or_404(EliminacionProtocolo, id=id)
 	instance.delete()
 	return redirect("bsadmin:l_eliminacionprotocolo")
 
 def activar_protocolo(request, id=None):
+	if not request.user.is_authenticated() or request.user.is_staff:
+		raise Http404
 	instance = get_object_or_404(DetalleAnalisisPadre, id=id)
 	instance.activo = True
 	instance.save()
@@ -542,6 +589,8 @@ def j_tercerizarid(request,id=None):
 	return JsonResponse(queryset, safe=False)
 
 def l_tercerizar(request):
+	if not request.user.is_authenticated() or request.user.is_staff:
+		raise Http404
 	queryset = Tercerizacion.objects.all().order_by('id')
 	query = request.GET.get("q")
 	if query:
@@ -567,6 +616,8 @@ def l_tercerizar(request):
 	return render(request, "list_tercerizar.html", context)
 
 def tercerizar(request, id=None):
+	if not request.user.is_authenticated() or request.user.is_staff:
+		raise Http404
 	#traigo info de toda la solicitud de analisis
 	instance = get_object_or_404(DetalleAnalisisPadre, id=id)
 	form = TercerizacionForm(request.POST or None)
@@ -601,6 +652,8 @@ def a_tercerizar(request):
 	return render(request, "alta_tercerizar.html", context)
 """	
 def v_tercerizar(request, id=None):
+	if not request.user.is_authenticated() or request.user.is_staff:
+		raise Http404
 	instance = get_object_or_404(Tercerizacion, id=id)
 	context = {
 		"instance" : instance,
@@ -610,6 +663,8 @@ def v_tercerizar(request, id=None):
 
 
 def u_tercerizar(request, idt=None, iddap=None):
+	if not request.user.is_authenticated() or request.user.is_staff:
+		raise Http404
 	instance_t = get_object_or_404(Tercerizacion,id=idt)
 	instance_dap = get_object_or_404(DetalleAnalisisPadre, id=iddap)
 	form = TercerizacionForm(request.POST or None, instance=instance_t)
@@ -628,6 +683,8 @@ def u_tercerizar(request, idt=None, iddap=None):
 	return render(request, "tercerizar.html", context)
 
 def hojadetrabajo(request, id=None):
+	if not request.user.is_authenticated() or request.user.is_staff:
+		raise Http404
 	#traigo info de toda la solicitud de analisis
 	instance = get_object_or_404(DetalleAnalisisPadre, id=id)
 	# traigo los parametros asociados al diagnostico
@@ -638,7 +695,11 @@ def hojadetrabajo(request, id=None):
 	all_indiv_de_solic = DetalleAnalisis.objects.distinct('individuoPadre').filter(solicitud=solic)
 	grupo_list_t = Parametros.objects.distinct('grupo').filter(diagnostico=diag, visualizacion1="T")
 	grupo_list_i = Parametros.objects.distinct('grupo').filter(diagnostico=diag, visualizacion1="I")
-	
+
+	# trae un unico individuo para la parte del listar los ítmes una sola vez en el HTML	
+	cant_indi = len(all_indiv_de_solic)
+	indi = all_indiv_de_solic[cant_indi-1]
+
 	# listado de valores de referencia filtrado por diag + param + especie
 	vdr_all = ValoresReferencia.objects.all()
 	vdr_list=[]
@@ -650,6 +711,8 @@ def hojadetrabajo(request, id=None):
 	#print("vdr_list")
 	#print(vdr_list)
 	
+	# listado de DetallesAnalisis como diccionario:
+	# {Key: DA, Value: valor}
 	da_all = DetalleAnalisis.objects.all().filter(solicitud=solic)
 	da_list = {}
 	for p in all_param_del_diag:
@@ -659,23 +722,24 @@ def hojadetrabajo(request, id=None):
 				if da.parametros == p:
 					if da.individuoPadre == i.individuoPadre:
 						if da.valor != '':
-							print("entro 1")
+							#print("entro 1")
 							valor=da.valor
 							#print(linea)
 							ban1=1
 						else:
 							for vdr in vdr_list:
 								if vdr.parametros == p:
-									print("entro 2")
+									#print("entro 2")
 									valor=vdr.valorDef
 									ban1=1
 							if ban1==0:
-								print("entro 3")
+								#print("entro 3")
 								valor=""
 								ban1=1
 					da_list[da]=valor
-	print(da_list)			
+	#print(da_list)
 
+	# diccionario de grupos_tabla separados por rangos de a 5 parametros
 	grupos = {}
 	for g in grupo_list_t:
 		cant=0
@@ -701,9 +765,10 @@ def hojadetrabajo(request, id=None):
 				ban=False
 		grupos[k] = list_r
 	#print(grupos)
+	
 	context = {
-		#"da_all": da_all,
 		"title" : "Registro de Resultados",
+		"indi": indi,
 		"da_list": da_list,
 		"instance" : instance,
 		"parametros_list" : all_param_del_diag,
